+++ src/Micro/Formations/SquareFormation.cpp
@@ -20,7 +20,7 @@ void SquareFormation::computeToPositions(const std::vector<pBayesianUnit>& vUnit
 	end_positions.clear();
     if (vUnits.size() == 1)
     {
-        end_positions.push_back(Position(center.x, center.y));
+        end_positions.push_back(Position((int)center.x, (int)center.y));
         return;
     }
     int maxDim = 0;
